requirements.txt
Django==4.2.14
django-ckeditor-5==0.2.13
django-jazzmin==3.0.0
django-leaflet==0.30.1
django-modeltranslation==0.19.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
geoip2==4.8.0
phonenumbers==8.13.40
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.6.0
djangorestframework-simplejwt==5.3.1
google-auth==2.32.0